Correct Precedence by changing the grammar.

+++ interpret.cpp
@@ -52,7 +52,7 @@ a_number operator/(a_number left , a_number right)
                 return std::get<double>(left) / std::get<double>(right);
 
         if(std::holds_alternative<int>(left)  && std::holds_alternative<int>(right))
-                return std::get<int>(left) / std::get<int>(right);
+		return ( static_cast<double>(std::get<int>(left)) / static_cast<double>(std::get<int>(right)) );
 
         if(std::holds_alternative<int>(left) && std::holds_alternative<double>(right))
                 return (static_cast<double>(std::get<int>(left))  / std::get<double>(right));
@@ -60,6 +60,7 @@ a_number operator/(a_number left , a_number right)
         if(std::holds_alternative<double>(left) && std::holds_alternative<int>(right))
                 return (std::get<double>(left) / static_cast<double>(std::get<int>(right)));
 }
+
 a_number interpret_ast(a_ast_node* ast)
 {
 	a_number leftval , rightval;
